Introduce `ptg` URL parameter for selecting parameter transform groups
The values of the `ptg` parameters will be matched with the group ids
in the `parameter-transforms.json` config file. The selected groups
will than be loaded into the app. Specify the parameter multiple times
to select several groups. If the parameter is not provided, all groups
will be loaded. Provide an unknown group id (e.g. '') to keep the list
of available parameter transforms empty.