Work around flow visualization issues
This introduces a simple configurable `--max-dashoffset-step` CSS
variable to provide a boundary for the dash offset step when animating
the flows. The flow edges use a repeating pattern. Therefore, large
enough flow values would make the flow vizualization look like the flow
goes backwards or alternates between forwards and backwards if no
boundary is used. The previous boundary computation was convoluted and
had issues when the line width went to 0. Hence, it needed to be
replaced with a simpler method, at least temporarily.